requirements.txt
django
djangorestframework
pandas
numpy
matplotlib
social-auth-app-django
social-auth-core
python-magic
scikits.bootstrap
python-magic-bin==0.4.14
python-magic==0.4.27
djangorestframework-simplejwt
django-allauth